+++ frontend/script/dashboard.js
@@ -61,9 +61,11 @@ function dropdown_menu() {
 }
 
 logout_btn.addEventListener("click", async () => {
-    // const response = await fetch(`${baseUrl}/users/logout`);
-    localStorage.clear();
-    window.location.href = "../index.html";
+    let confirmation = confirm("Are you sure you want to log out?");
+    if (confirmation) {
+        localStorage.clear();
+        window.location.href = "../index.html";
+    }
 })
 
 delete_btn.addEventListener("click", async () => {
@@ -109,6 +111,7 @@ edit_btn.addEventListener("click", async () => {
     let submit_btn = document.getElementById("submit_btn");
     submit_btn.addEventListener("click", async(e) => {
         e.preventDefault();
+        submit_btn.innerHTML = `<i class="fa fa-spinner fa-spin"></i>`;
         // console.log("Please enter your new password")
         let new_password = document.getElementById("new-password").value;
         let confirm_password = document.getElementById("confirm-password").value;
@@ -130,6 +133,7 @@ edit_btn.addEventListener("click", async () => {
             let response = await request.json();
             // console.log(response);
             alert(response.msg);
+            window.location.reload();
         }
         else{
             console.log("false");
@@ -147,7 +151,7 @@ function displayStats(userInfo) {
     let platforms = [], platforms_wise_clicks = [];
     let locations = [], locations_wise_clicks = [];
     let browsers = [], browsers_wise_clicks = [];
-    let createdAt = [];
+    // let createdAt = [];
 
 
     for (let i = 0; i < userInfo.date.length; i++) { date.push(userInfo.date[i]._id); date_wise_clicks.push(userInfo.date[i].count); }
